Use custom exceptions instead of runtime_error
Use custom exceptions LocalException and RethrownException instead of std::runtime_error to make it possible for unit tests to see if the exception comes from this program or external code. RethrownException is used when throwing as a reaction to exception from external code and LocalException when exception means something else. Purpose of this is to spot uncaught exceptions.